requirements.txt
django>=1.11,<2.0 #LTS
django-extensions==2.1.4
django-filter==2.0.0
djangorestframework==3.9.0
gunicorn==19.9.0
psycopg2==2.7.6.1
pylama==7.6.6
pytest-django==3.4.5
pytest==4.1.1
django-dbbackup==3.2.0
boto3==1.9.79
dj-database-url==0.5.0
whitenoise==4.1.2